# move simple functions to static inline in model_help.[ch]