--- a/SimulationRuntime/c/simulation/solver/model_help.h
+++ b/SimulationRuntime/c/simulation/solver/model_help.h
@@ -150,18 +150,44 @@ modelica_real _event_div_real(modelica_real x1, modelica_real x2, modelica_integ
 /* functions used for relation which
  * are not used as zero-crossings
  */
-modelica_boolean Less(double a, double b);
-modelica_boolean LessEq(double a, double b);
-modelica_boolean Greater(double a, double b);
-modelica_boolean GreaterEq(double a, double b);
+static OMC_INLINE modelica_boolean Less(double a, double b)
+{
+  return a < b;
+}
+
+static OMC_INLINE modelica_boolean LessEq(double a, double b)
+{
+  return a <= b;
+}
+
+static OMC_INLINE modelica_boolean Greater(double a, double b)
+{
+  return a > b;
+}
+
+static OMC_INLINE modelica_boolean GreaterEq(double a, double b)
+{
+  return a >= b;
+}
+
+
+
 
 /* functions used to evaluate relation in
  * zero-crossing with hysteresis effect
  */
 modelica_boolean LessZC(double a, double b, modelica_boolean);
-modelica_boolean LessEqZC(double a, double b, modelica_boolean);
 modelica_boolean GreaterZC(double a, double b, modelica_boolean);
-modelica_boolean GreaterEqZC(double a, double b, modelica_boolean);
+static OMC_INLINE modelica_boolean GreaterEqZC(double a, double b, modelica_boolean direction)
+{
+  return !LessZC(a, b, !direction);
+}
+
+static OMC_INLINE modelica_boolean LessEqZC(double a, double b, modelica_boolean direction)
+{
+  return !GreaterZC(a, b, !direction);
+}
+
 
 extern int measure_time_flag;