avoids unnecessary function calls and could help compiler to optimize relations