@hkiel 2017-04-24
Can you please restart ScalabeTestSuite once this is PR merged?