@sjoelund 2017-05-05
I guess this should be benchmarked since we compile code with -O0 by default. Or maybe we should change it to -O1 or -Og so inlining, etc works as expected...