@hkiel 2017-05-05
@sjoelund Where is the default optimization level set? I cannot find "-O0" in *.mo files.