@hkiel 2017-05-05
@sjoelund @casella
Ok, I use `setCFlags("-O1")` now and "-n=1" for omc.
I simulated `ScalableTestSuite.Electrical.BreakerCircuits.ScaledExperiments.BreakerNetworkDelayed_N_320` and  compile time went up from 10s -> 37s, while simulation time went down from 65s -> 45s.
When using "-Os" compile time eveng goes up to 55s and simulation time down to 34s.

An impact of this modification could not be seen.

I tested only one model, but overall I guess we have a slower system when C-compiler is optimizing.
Maybe users can switch on optimization for really slow simulations (there are some in the ScalableTestSuite)...