@sjoelund 2017-05-05
Yes, users can do turn on the optimiser flags. But in general in most people don't, even when benchmarking OpenModelica simulations...