@hkiel 2017-05-06
Can we emit a hint to the user that turning on optimization can speed up things?
Maybe that can be shown when simulation takes more than a few seconds.