@hkiel 2017-05-13
I'm closing this, because C-compiler does not take advantage of this when not optimizing, and when optimizing, compilation slow down is larger than speed up of simulation run.